REST endpoint to toggle a feature in one or more environments
TODO:
- [ ] Audit logging
- [ ] Documentation
