+++ packages/back-end/src/api/features/features.router.ts
@@ -1,0 +1,12 @@
+import { Router } from "express";
+import { listFeatures } from "./listFeatures";
+import { toggleFeature } from "./toggleFeature";
+
+const router = Router();
+
+// Feature Endpoints
+// Mounted at /api/v1/features
+router.get("/", listFeatures);
+router.post("/:key/toggle", toggleFeature);
+
+export default router;
